requirements.txt
requests
pytest
pyyaml
pandas
numpy
pytest-html
allure-pytest